% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_dimensions.r
\name{dataset_dimensions}
\alias{dataset_dimensions}
\title{Return the dimensions of a dataset}
\usage{
dataset_dimensions(dataset)
}
\arguments{
\item{dataset}{The last part of the URI of a dataset}
}
\value{
The dimensions of the dataset at
\code{statistics.gov.scot/data/<dataset>}
}
\description{
Return the dimensions of a dataset
}
\examples{
scotgov:::dataset_dimensions("average-household-size")
}
