% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{Datasets}
\usage{
datasets(search = "")
}
\arguments{
\item{search}{Case-sensitive search term to fitler datasets. By default will return all
datasets.}
}
\value{
dataframe of available datasets (matching search term if provided).
}
\description{
Datasets
}
\examples{
data <- datasets(search = "bicycle") # search datasets
data <- datasets() # return all datasets
}
