% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_lookup}
\alias{geo_lookup}
\title{Geographic Lookup Data}
\format{
A data frame with 53940 rows and 10 variables:
\itemize{
  \item{\code{datazone_2001}}{ - 2001 data zone code}
  \item{\code{datazone_2011}}{ - 2011 data zone code}
}
}
\source{
\emph{Scottish Government,} 2015. \href{https://www2.gov.scot/Topics/Statistics/sns/SNSRef/DZMatchingFile}{Data Zone Matching File (Area and Pop. based 2001 - 2011)}
}
\usage{
geo_lookup
}
\description{
A simple lookup data set linking data zones to high-level
  geographies across Scotland. The included geographies account for 2001 and
  2011 data zones.
}
\details{
Some detail
}
\keyword{datasets}
