% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(search = "", refresh = FALSE, ask = TRUE)
}
\arguments{
\item{search}{Search term}

\item{refresh}{Refresh cached data. If data has changed remotely, use this to
update or renew corrupted data/cache. This will download data again and
update cache.}

\item{ask}{If FALSE, user will not be prompted for input to download data.
This is useful to automate data downloads and cache updated or speed up the
process if downloading many datasets at once.}
}
\description{
Search and return matching datasets from https://opendata.scot/ API.
}
