% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sg_datasets.r
\name{list_sg_datasets}
\alias{list_sg_datasets}
\title{Return a \code{data.frame} of all data sets available on statistics.gov.scot}
\usage{
list_sg_datasets()
}
\value{
A \code{data.frame} of all data sets available on statistics.gov.scot
}
\description{
Return a \code{data.frame} of all data sets available on statistics.gov.scot
}
\examples{
scotgov:::list_sg_datasets()
}
