% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_all_datasets.R
\name{ods_all_datasets}
\alias{ods_all_datasets}
\title{find all datasets in opendatascotland}
\usage{
ods_all_datasets()
}
\value{
A list of data from statistics.gov.scot
}
\description{
\code{ods_all_datasets} returns a list of all datasets from statistics.gov.scot
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
ods_all_datasets()

}
