% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_label_uri.R
\name{ods_create_label_lookup}
\alias{ods_create_label_lookup}
\title{functions to label a dataset full of uri}
\usage{
ods_create_label_lookup(uri_list)
}
\value{
\code{tibble}.
}
\description{
\code{ods_create_label_lookup} returns a dataset that is then used as a lookup
}
