% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_find_higher_geographies.R
\name{ods_find_higher_geographies}
\alias{ods_find_higher_geographies}
\title{Find all geographies that contain, a certain geography}
\usage{
ods_find_higher_geographies(geography)
}
\arguments{
\item{geography}{\code{string}. A valid geography unit from \href{https://statistics.gov.scot/}{statistics.gov.scot}.}
}
\value{
\code{tibble}.
}
\description{
\code{ods_find_higher_geographies} returns a dataset from \href{https://statistics.gov.scot/}{statistics.gov.scot} of all geographies which contain the geography provided
}
\details{
The \code{geography} parameter must be passed a valid geography unit.
}
\examples{
ods_find_higher_geographies("S02001257")

}
