% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_get_csv.R
\name{ods_get_csv}
\alias{ods_get_csv}
\title{Download a dataset csv}
\usage{
ods_get_csv(dataset)
}
\arguments{
\item{dataset}{\code{string}. The identifying final part of a URI for a dataset on \href{https://statistics.gov.scot/}{statistics.gov.scot}.}
}
\value{
\code{tibble}.
When invalid arguments are used returns \code{NULL} with \code{warning} (see \code{ods_error_message}).
}
\description{
\code{ods_get_csv} returns a dataset from \href{https://statistics.gov.scot/}{statistics.gov.scot}
The \code{dataset} parameter must be passed a valid dataset name (a full list can be obtained by calling \code{ods_all_datasets()}.
}
\examples{
ods_get_csv("average-household-size")

}
