% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_get_geoshape.R
\name{ods_get_geoshape}
\alias{ods_get_geoshape}
\title{Return a polygon of a valid statistical geography}
\usage{
ods_get_geoshape(geography)
}
\arguments{
\item{geography}{\code{string}. A valid geography unit from \href{https://statistics.gov.scot/}{statistics.gov.scot}.}
}
\value{
\code{string}.
}
\description{
\code{ods_get_geoshape} returns a string from \href{https://statistics.gov.scot/}{statistics.gov.scot} of all geographies which contain the geography provided
}
\details{
The \code{geography} parameter must be passed a valid geography unit.
}
\examples{
ods_get_geoshape("S02001257")

}
