% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_label_uri.R
\name{ods_label_uri}
\alias{ods_label_uri}
\title{Find all geographies that contain, or are contained by, a certain geography}
\usage{
ods_label_uri(dataset, uri)
}
\value{
\code{tibble}.
}
\description{
\code{ods_label_uri} labels a dataset that comprises of uri
}
