% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_metadata.R
\name{ods_metadata}
\alias{ods_metadata}
\title{Return the metadata used by a dataset}
\usage{
ods_metadata(dataset)
}
\arguments{
\item{dataset}{\code{string}. The identifying final part of a URI for a dataset on \href{https://statistics.gov.scot/}{statistics.gov.scot}.}
}
\value{
\code{data.frame}. All schemes used by the dataset named in the dataset argument.
}
\description{
\code{ods_metadata} returns all metadata used by a dataset on \href{https://statistics.gov.scot/}{statistics.gov.scot}
}
\details{
The \code{dataset} parameter must be passed a valid dataset name (a full list can be obtained by calling \code{ods_all_datasets()}).
}
\examples{
ods_metadata("average-household-size")

}
