% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_structure.R
\name{ods_structure}
\alias{ods_structure}
\title{Return a list of the structure of a dataset}
\usage{
ods_structure(dataset)
}
\arguments{
\item{dataset}{\code{string}. The identifying final part of a URI for a dataset on \href{https://statistics.gov.scot/}{statistics.gov.scot}.}
}
\value{
\code{list}. Returns all schemes and concepts within a dataset.
}
\description{
\code{ods_structure} returns a list of data about a dataset on \href{https://statistics.gov.scot/}{statistics.gov.scot}
}
\details{
The \code{dataset} parameter must be passed a valid dataset name (a full list can be obtained by calling \code{ods_all_datasets()}.
}
\examples{
ods_structure("average-household-size")

}
