% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotgov_get.R
\name{scotgov_get}
\alias{scotgov_get}
\title{Download data from statistics.gov.scot}
\usage{
scotgov_get(dataset)
}
\arguments{
\item{endpoint}{An API endpoint for statistics.gov.scot}
}
\value{
A tibble of data from statistics.gov.scot
}
\description{
\code{scotgov_get} returns a tibble of data from statistics.gov.scot
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
scotgov_get("http://statistics.gov.scot/sparql")

}
