% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read_query_file.R
\name{read_query_file}
\alias{read_query_file}
\title{Read SPARQL Query File}
\usage{
read_query_file(file_name)
}
\arguments{
\item{file_name}{Name of SPARQL file to source.}
}
\value{
A character scalar.
}
\description{
The function searches package content and selects \code{sparql}
  file matching \code{x}. The file name can be passed with or without
  \code{sparql} extension.
}
\details{
The function is used internally and it's not exported. In addition
  to getting the file name the function does minor pre-processing on the file
  like removing the comments.
}
\examples{
read_query_file(x = "find_lower_level_geographies")
}

