% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_query.r
\name{get_dataset_query}
\alias{get_dataset_query}
\title{Generate SPARQL query to download data from statistics.gov.scot}
\usage{
get_dataset_query(dataset, start_date = NULL, end_date = NULL,
  geography = NULL)
}
\arguments{
\item{endpoint}{An API endpoint for statistics.gov.scot}
}
\value{
A string consisting of a valid SPARQL query
}
\description{
\code{get_dataset_query} returns a tibble of data from statistics.gov.scot
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
get_dataset_query("average-house-size")

}
