% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ods.R
\name{get_ods}
\alias{get_ods}
\title{Get Data}
\usage{
get_ods(data = NULL, refresh = FALSE, ask = TRUE)
}
\arguments{
\item{data}{Dataframe from \code{search_ods()} or if default (NULL), will download
all datasets.}

\item{refresh}{Refresh cached data. If data has changed remotely, use this to
update or renew corrupted data/cache. This will download data again and
update cache.}

\item{ask}{If FALSE, user will not be prompted for input to download data.
This is useful to automate data downloads and cache updated or speed up the
process if downloading many datasets at once.}
}
\description{
Return matching data from https://opendata.scot/ API.
}
