% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sg_datasets.r
\name{list_sg_datasets}
\alias{list_sg_datasets}
\title{find all datasets in opendatascotland}
\usage{
list_sg_datasets()
}
\value{
A list of data from statistics.gov.scot
}
\description{
\code{list_sg_datasets} returns a list of all datasets from statistics.gov.scot
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
list_sg_datasets()

}
