% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_all_datasets.R
\name{ods_all_datasets}
\alias{ods_all_datasets}
\title{List all datasets}
\usage{
ods_all_datasets()
}
\value{
\code{data.frame}. Contains the URI, Name, Publisher and Dataset Name of every dataset.
}
\description{
\code{ods_all_datasets} returns details of all datasets available from \href{https://statistics.gov.scot/}{statistics.gov.scot}
}
\examples{
ods_all_datasets()

}
