% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_dataset.R
\name{ods_dataset}
\alias{ods_dataset}
\title{Download data from statistics.gov.scot}
\usage{
ods_dataset(dataset, start_date = NULL, end_date = NULL,
  geography = NULL, ...)
}
\arguments{
\item{dataset}{The last part of the URI for a dataset on statistics.gov.scot}

\item{start_date}{Filter data points on or after this date}

\item{end_date}{Filter data points after this date}

\item{geography}{An S code - filter data points within this geography}

\item{...}{arbitrary filters requiring valid scheme = value structure}
}
\value{
A \code{data.frame} of data from statistics.gov.scot
}
\description{
\code{ods_dataset} returns a tibble of data from statistics.gov.scot
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
ods_dataset("average-household-size")

}
