% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_print_query.R
\name{ods_print_query}
\alias{ods_print_query}
\title{Generate SPARQL query to download data from statistics.gov.scot}
\usage{
ods_print_query(dataset, start_date = NULL, end_date = NULL,
  geography = NULL, ...)
}
\arguments{
\item{dataset}{An API endpoint for statistics.gov.scot}

\item{start_date}{A date}

\item{end_date}{A date}

\item{geography}{A string representing a valid SG geography code, or portion thereof}

\item{...}{arbitrary filters requiring valid scheme = value structure}
}
\value{
A string consisting of a valid SPARQL query
}
\description{
\code{ods_print_query(} returns a tibble of data from statistics.gov.scot
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
ods_print_query("average-house-size")

}
