% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_print_query.R
\name{ods_print_query}
\alias{ods_print_query}
\title{Generate a SPARQL query to call on API}
\usage{
ods_print_query(dataset, geography = NULL, ...)
}
\arguments{
\item{dataset}{\code{string}. The identifying final part of a URI for a dataset on \href{https://statistics.gov.scot/}{statistics.gov.scot}.}

\item{geography}{\code{string}. A valid SG geography code, or portion thereof.}

\item{...}{Arbitrary filters requiring valid scheme = value structure.}
}
\value{
\code{string}. A valid SPARQL query.
}
\description{
\code{ods_print_query} returns a valid SPARQL query to call on the \href{https://statistics.gov.scot/}{statistics.gov.scot} API
}
\details{
The \code{dataset} parameter must be passed a valid dataset name (a full list can be obtained by calling \code{ods_all_datasets()}.
Other parameters can also be used to filter the dataset.
}
\examples{
ods_print_query("average-household-size")

}
