% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_schemes.R
\name{ods_schemes}
\alias{ods_schemes}
\title{Return the schemes of a dataset}
\usage{
ods_schemes(dataset)
}
\arguments{
\item{dataset}{The last part of the URI of a dataset}
}
\value{
The schemes of the dataset at
\code{statistics.gov.scot/data/<dataset>} as a data frame
}
\description{
Return the schemes of a dataset
}
\examples{
opendatascot:::ods_schemes("average-household-size")
}
