% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods-search.R
\name{ods_search}
\alias{ods_search}
\title{Search Open Data Scot Datasets}
\usage{
ods_search(search = "")
}
\arguments{
\item{search}{A vector of search terms. Searches are performed on the title
of the datasets and are case insensitive. If no search terms provided,
metadata for all datasets will be returned.}
}
\value{
dataframe of metadata for available datasets (matching search term if
provided).
}
\description{
This function will return metadata associated with datasets available on
opendata.scot
}
\examples{
all_datasets <- ods_search() # return ALL datasets
data <- ods_search(search = "bicycle") # search datasets
data <- ods_search(search = c("bicycle","bins")) # multiple search terms
}
