% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_structure.R
\name{ods_structure}
\alias{ods_structure}
\title{View structure of data from statistics.gov.scot}
\usage{
ods_structure(dataset)
}
\arguments{
\item{dataset}{An API endpoint for statistics.gov.scot that corresponds to a dataset}
}
\value{
A list of data from statistics.gov.scot
}
\description{
\code{ods_structure} returns a list of data about a dataset from statistics.gov.scot
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
ods_structure("average-household-size")

}
