% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotgov_get.R
\name{scotgov_get}
\alias{scotgov_get}
\title{Download data from statistics.gov.scot}
\usage{
scotgov_get(dataset, start_date = NULL, end_date = NULL,
  geography = NULL)
}
\arguments{
\item{dataset}{The last part of the URI for a dataset on statistics.gov.scot}

\item{start_date}{Filter data points on or after this date}

\item{end_date}{Filter data points after this date}

\item{geography}{An S code - filter data points within this geography}
  
\item{...}{arbitrary filters requiring valid dimension = value structure}
}
\value{
A \code{data.frame} of data from statistics.gov.scot
}
\description{
\code{scotgov_get} returns a tibble of data from statistics.gov.scot
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
scotgov_get("average-household-size")

}
