% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-ods.R
\name{search_ods}
\alias{search_ods}
\title{Datasets}
\usage{
search_ods(search = "")
}
\arguments{
\item{search}{A case insensitive search term to fitler datasets. By default
will return all datasets.}
}
\value{
dataframe of available datasets (matching search term if provided).
}
\description{
Datasets
}
\examples{
data <- search_ods(search = "bicycle") # search datasets
data <- search_ods() # return all datasets
}
